rename ONLY the 4 vsthemes
DPS themes are back to their default names.

The *.vsthemes provided on the file system get named in memory to *NG

If we don't find the the theme name pulled from settings, blank it out. This should ONLY happen after upgrading from previous releases for users of the 4 *.vsthemes providedon the file system:  "darcula", "vs2015blue", "vs2015dark" , "vs2015light"

The 3 vs2015* ones are NOT to be confsed with the DPS provided versions of the almost identical name.

Jumping through hoops rather than renaming files/theme names to avoid unexpected theme changes for as many users as possible.