Increased VNC test connection timeout from 150ms to 500ms
It has been found at times that when connecting to a VNC host that the connection times-out even though the host is powered on. This happens mostly when you are connected to the network through WiFi. The time out value for the VNC test connection has been increased from 150ms to 500ms.